+++ CHANGELOG.md
@@ -1,8 +1,19 @@
+*Unreleased*
+
+Dependency upgrades, logging and backup/recovery improvements.
+## v0.2.0
+### Breaking changes
+* [68](https://github.com/crypto-com/tmkms-light/pull/68) production logging in NE
+* [55](https://github.com/crypto-com/tmkms-light/pull/55) SGX init and recovery commands paramater change
+* [80](https://github.com/crypto-com/tmkms-light/pull/80) SGX backup and recovery in cloud environments
+
 *March 18, 2021*
+
 Tendermint-rs and prost dependencies upgraded.
 ## v0.1.2
 
 *March 16, 2021*
+
 The same as the initial released version, but with a few dependency upgrades
 ## v0.1.1